requirements.txt
arabic-reshaper==2.1.4
arrow==1.2.3
asgiref==3.5.2
asn1crypto==1.5.1
Babel==2.11.0
binaryornot==0.4.4
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
click==8.1.3
cookiecutter==2.1.1
cryptography==38.0.4
cssselect2==0.7.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==1.2.0
Django==4.1.3
django-appconf==1.0.5
django-browser-reload==1.6.0
django-compressor==4.1
django-easy-pdf3==0.1.4
django-environ==0.9.0
django-import-export==3.0.1
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
django-tailwind==3.4.0
et-xmlfile==1.1.0
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
idna==3.4
Jinja2==3.1.2
jinja2-time==0.2.0
lxml==4.9.2
MarkupPy==1.14
MarkupSafe==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
oscrypto==1.3.0
phonenumbers==8.13.0
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
pyHanko==0.15.1
pyhanko-certvalidator==0.19.7
PyPDF3==1.0.6
python-bidi==0.4.2
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
rcssmin==1.1.0
reportlab==3.6.12
requests==2.28.1
rjsmin==1.2.0
six==1.16.0
sqlparse==0.4.3
svglib==1.4.1
tablib==3.2.1
text-unidecode==1.3
tinycss2==1.2.1
tqdm==4.64.1
tzdata==2022.7
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
webencodings==0.5.1
whitenoise==6.2.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
